pyproject.toml
[project]
name = "bedrock-mantle"
version = "0.1.0"
description = "CLI for exploring Amazon Bedrock OpenAI-compatible APIs (Project Mantle)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.59.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
bedrock-mantle = "main:cli"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-timeout>=2.3.0",
    "ruff>=0.8.0",
    "vulture>=2.14",
    "pre-commit>=4.0.0",
    "mypy>=1.14.0",
    "types-click>=7.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"

[tool.vulture]
min_confidence = 80
paths = ["main.py", "tests/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (require real API credentials)",
    "slow: marks tests as slow (may take longer due to API calls)",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = false
warn_unused_ignores = true
ignore_missing_imports = true
disable_error_code = ["arg-type", "call-overload", "union-attr", "dict-item", "assignment"]
